pyproject.toml
[project]
name = "bad-apple-attempt-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.16",
    "databento>=0.68.1",
    "exchange-calendars>=4.11.3",
    "lseg-data>=2.1.1",
    "modal>=1.3.0.post1",
    "numba>=0.63.1",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pyarrow>=22.0.0",
    "scipy>=1.16.3",
    "tqdm>=4.67.1",
    "yfinance>=1.0",
    "yt-dlp>=2025.12.8",
]
